% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission_helper_functions.R
\name{build_rds}
\alias{build_rds}
\title{Build RDS Submission Object}
\usage{
build_rds(file, session, is_test = FALSE)
}
\arguments{
\item{file}{location to save RDS file}

\item{session}{session object from shiny with learnr}

\item{is_test}{check if testing function}
}
\value{

}
\description{
Build RDS Submission Object
}
