% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission_aggregation.R
\name{create_submission_aggregation}
\alias{create_submission_aggregation}
\title{Create Submission Aggregation}
\usage{
create_submission_aggregation(paths, tutorial_id, new_dir)
}
\arguments{
\item{paths}{local paths to saved rds files}

\item{tutorial_id}{id of tutorial to save and process}

\item{new_dir}{new gdrive folder created in gdrive_access.R}
}
\value{

}
\description{
Create Submission Aggregation
}
