% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission_functions.R
\name{encode_obj}
\alias{encode_obj}
\title{Encode an R object into hashed text; from github::rundel/learnrhash}
\usage{
encode_obj(obj, compress = c("bzip2", "gzip", "xz", "none"))
}
\arguments{
\item{obj}{R object}

\item{compress}{Compression method.}
}
\description{
Encode an R object into hashed text; from github::rundel/learnrhash
}
