% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdrive_access.R
\name{gdrive_access}
\alias{gdrive_access}
\title{Google Drive Submission Processing}
\usage{
gdrive_access(tutorial_id, year, month, day, temp_path, rds_paths)
}
\arguments{
\item{tutorial_id}{id of tutorial to save and process}

\item{year}{year from parsed_date}

\item{month}{month from parsed_date}

\item{day}{day from parsed_date}

\item{temp_path}{path of tempdir()}

\item{rds_paths}{paths of saved rds files from gmail}
}
\value{
folder name of submission results
}
\description{
Google Drive Submission Processing
}
