% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission_helper_functions.R
\name{get_label_list}
\alias{get_label_list}
\title{Get Label Order}
\usage{
get_label_list(sess)
}
\arguments{
\item{sess}{session object from shiny with learnr}
}
\value{
list of labels in the correct order
}
\description{
Get Label Order
}
