% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_access.R
\name{gmail_access}
\alias{gmail_access}
\title{Gmail Access and Submission Processing}
\usage{
gmail_access(query, key, secret)
}
\arguments{
\item{query}{used to filter mails to find submissions}

\item{key}{client id obtained from oauth credentials}

\item{secret}{client secret obtained from oauth credentials}
}
\value{
rds_paths paths of locally saved rds files from gmail
}
\description{
Gmail Access and Submission Processing
}
