% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_labels.R
\name{make_labels}
\alias{make_labels}
\title{Convert Names to Tribble Column Labels}
\usage{
make_labels(x)
}
\arguments{
\item{x}{A character vector of column names.}
}
\value{
A character vector of formatted column labels like \verb{~\\}name\``.
}
\description{
Converts a character vector of column names into a vector formatted as
\verb{~\\}name\`` strings for use as column headers in \code{tibble::tribble()}.
}
\author{
David Kane, Aashna Patel
}
