% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_exercise.R
\name{make_new_exercise}
\alias{make_new_exercise}
\title{Make New Exercise}
\usage{
make_new_exercise()
}
\value{
exercise skeleton with appropriate labels and numbers
}
\description{
Make New Exercise
}
