% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_q_with_answer.R
\name{make_new_q_with_answer}
\alias{make_new_q_with_answer}
\title{Tutorial Written Exercise (With Answer)}
\usage{
make_new_q_with_answer()
}
\value{
question skeleton with appropriate labels and numbers
}
\description{
An add-in for writingall.primer.tutorials.

It reads the latest exercise and section name
and then adds a question skeleton (with answers and no retries).

This should make things easier for tutorial-writers
because now a fast click can create most of the exercise for you.

There is also no need to keep track of the exercise numbers
because it is done for you in the add-in.

Keyboard shortcut should be Cmd + SHIFT + E (MAC)
}
