% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_p_table.R
\name{make_p_table}
\alias{make_p_table}
\title{Insert a standardized Preceptor or Population Table template}
\usage{
make_p_table(table = "preceptor", type = "predictive")
}
\arguments{
\item{table}{Character. Either "preceptor" or "population" (partial match allowed).}

\item{type}{Character. Either "predictive" or "causal" (partial match allowed).}
}
\description{
Insert a standardized Preceptor or Population Table template
}
