% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_rstudio_settings.R
\name{prep_rstudio_settings}
\alias{prep_rstudio_settings}
\title{Set Up RStudio for Student Use}
\usage{
prep_rstudio_settings()
}
\description{
This functions selects RStudio settings which make learning easier. It also
sets the "pkgType" global option to "binary" in your .Rprofile. Students,
especially those on Windows, should never install from source. Doing so fails
too often, and too confusingly.

You can examine your .Rprofile to confirm this change with
usethis::edit_r_profile()
}
