% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer.tutorials-package.R
\docType{package}
\name{primer.tutorials-package}
\alias{primer.tutorials-package}
\alias{primer.tutorials}
\title{primer.tutorials: Interactive Tutorials for Preceptor's Primer for Bayesian Data Science}
\description{
This package provides interactive tutorials that accompany "Preceptor's Primer
for Bayesian Data Science: Using the Cardinal Virtues for Inference". The
tutorials cover the fundamental concepts of data science through the lens of
Bayesian thinking, including data visualization, wrangling, modeling, and
statistical inference using modern R tools.
}
\section{Getting Started}{

To see available tutorials, run:
\code{learnr::available_tutorials("primer.tutorials")}

To start a tutorial, run:
\code{learnr::run_tutorial("141-stops", package = "primer.tutorials")}
}

\section{About the Primer}{

These tutorials are designed to complement "Preceptor's Primer for Bayesian
Data Science: Using the Cardinal Virtues for Inference", which teaches data
science through the framework of the four cardinal virtues: Wisdom, Justice,
Courage, and Temperance. The textbook is available
at \url{https://ppbds.github.io/primer/}.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PPBDS/primer}
  \item \url{https://ppbds.github.io/primer/tutorials/}
  \item Report bugs at \url{https://github.com/PPBDS/primer/issues}
}

}
\author{
\strong{Maintainer}: David Kane \email{dave.kane@gmail.com} (\href{https://orcid.org/0000-0002-6660-3934}{ORCID})

}
\keyword{internal}
