% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submission_functions.R
\docType{data}
\name{submission_server}
\alias{submission_server}
\alias{submission_ui}
\title{Tutorial submission functions}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
submission_server(input, output)

submission_ui
}
\arguments{
\item{input}{unused}

\item{output}{unused}
}
\description{
The following function was modified from Colin Rundel's learnrhash package,
available at https://github.com/rundel/learnrhash. Many thanks to Professor
Rundel, who has developed a fantastic tool for courses that teach R and use
the learnr package.

This note is also modified from Professor Rundel's description: Note that
when including these functions in a learnr Rmd document it is necessary that
the server function, \code{submission_server()}, be included in an R chunk where
\code{context="server"}. Conversely, any of the ui functions, \verb{*_ui()}, must \emph{not}
be included in an R chunk with a \code{context}.
}
\keyword{datasets}
