% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_current_tutorial.R
\name{check_current_tutorial}
\alias{check_current_tutorial}
\title{Check Current Tutorial}
\usage{
check_current_tutorial()
}
\value{
current document OVERWRITTEN and formatted with correct chunk labels
}
\description{
An add-in for formatting all.primer.tutorials.

It uses format_tutorial() to format the tutorial Rmd open
in the current editor
}
