% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tibble_from_submissions.R
\name{create_tibble_from_submissions}
\alias{create_tibble_from_submissions}
\title{Create Ordered Tibble from Submissions}
\usage{
create_tibble_from_submissions(objs, label_list)
}
\arguments{
\item{objs}{learnr session submissions}

\item{label_list}{order of code chunks}
}
\value{
tibble with ordered answers based on label_list
}
\description{
Create Ordered Tibble from Submissions
}
