% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label_list.R
\name{get_label_list}
\alias{get_label_list}
\title{Get Label Order}
\usage{
get_label_list(sess, is_test = FALSE)
}
\arguments{
\item{sess}{session object from shiny with learnr}

\item{is_test}{check if testing function}
}
\value{
list of labels in the correct order
}
\description{
Get Label Order
}
