% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_exercise.R
\name{make_new_exercise}
\alias{make_new_exercise}
\title{Tutorial Code Exercise}
\usage{
make_new_exercise()
}
\value{
exercise skeleton with appropriate labels and numbers
}
\description{
An add-in for writing all.primer.tutorials.

It reads the latest exercise and section name
and then adds an exercise skeleton.

This should make things easier for tutorial-writers
because now a fast click can create most of the exercise for you.

There is also no need to keep track of the exercise numbers
because it is done for you in the add-in.

Keyboard shortcut should be Cmd + SHIFT + E (MAC)
}
