% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_p_tables.R
\name{make_p_tables}
\alias{make_p_tables}
\title{Insert a standardized Preceptor or Population Table template}
\usage{
make_p_tables(
  is_causal = TRUE,
  unit_label = "Unit",
  outcome_label = if (is_causal) "Potential Outcomes" else "Outcome",
  treatment_label = "Treatment",
  covariate_label = "Covariates"
)
}
\arguments{
\item{is_causal}{Logical. If TRUE, generates template for causal analysis; else predictive.}

\item{unit_label}{Label (character) for the Unit/Time spanner}

\item{outcome_label}{Label (character) for the Outcome or Potential Outcomes spanner}

\item{treatment_label}{Label (character) for the Treatment spanner (only used if is_causal = TRUE)}

\item{covariate_label}{Label (character) for the Covariates spanner}
}
\description{
Insert editable templates for Preceptor and Population tables.
}
\details{
This function creates a five-chunk Quarto-ready template for Preceptor and Population Tables.
The output is inserted directly into the current document at the cursor.
}
