% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.primer.tutorials-package.R
\docType{package}
\name{primer.tutorials-package}
\alias{primer.tutorials}
\alias{primer.tutorials-package}
\title{primer.tutorials: Tutorials for Preceptor's Primer for Bayesian Data Science}
\description{
Collection of tutorials for use with Preceptor's Primer for Bayesian Data Science. Makes extensive use of the tools in tutorial.helpers package. This collection is currently in flux as we move much of the content to the new r4ds.tutorials package. Several of the remaining tutorials, especially for after Chapter 6, are not reliable.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ppbds.github.io/primer.tutorials/}
  \item Report bugs at \url{https://github.com/ppbds/primer.tutorials/issues}
}

}
\author{
\strong{Maintainer}: David Kane \email{dave.kane@gmail.com} (\href{https://orcid.org/0000-0002-6660-3934}{ORCID})

Authors:
\itemize{
  \item Becca Gill \email{beccabgill116@gmail.com} (\href{https://orcid.org/0000-0003-4084-4868}{ORCID})
  \item Miriam Heiss \email{memerheiss@gmail.com}
}

}
\keyword{internal}
