% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer.tutorials-package.R
\docType{package}
\name{primer.tutorials-package}
\alias{primer.tutorials}
\alias{primer.tutorials-package}
\title{primer.tutorials: Tutorials for Preceptor's Primer for Bayesian Data Science}
\description{
Collection of tutorials for use with Preceptor's Primer for Bayesian Data Science. Makes extensive use of the tools in 'tutorial.helpers' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ppbds.github.io/primer.tutorials/}
  \item Report bugs at \url{https://github.com/ppbds/primer.tutorials/issues}
}

}
\author{
\strong{Maintainer}: David Kane \email{dave.kane@gmail.com} (\href{https://orcid.org/0000-0002-6660-3934}{ORCID})

}
\keyword{internal}
