% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input_tribble.R
\name{write_input_tribble}
\alias{write_input_tribble}
\title{Write a tribble template with placeholders for a character vector of column names}
\usage{
write_input_tribble(names)
}
\arguments{
\item{names}{Character vector of column names.}
}
\value{
Character string containing the R code for an input tribble with placeholders \code{"..."},
formatted for manual editing with aligned columns.
}
\description{
Generates a character string representing an R tibble::tribble()
with the specified column names and 3 rows filled with placeholder
text \code{"..."}. Column values are aligned under their headers for easy editing.
}
\examples{
write_input_tribble(c("Unit", "Year", "Outcome", "Treatment"))

}
